requirements.txt
Pillow==8.4.0
numpy
diffusers
opencv-python
transformers
peft
accelerate
torch==2.4.1+cu124
torchvision==0.19.1+cu124
--index-url https://download.pytorch.org/whl/cu124
